+++ examples/nextjs-siwe/src/pages/_app.tsx
@@ -2,11 +2,11 @@ import '@/styles/globals.css';
 import { siweClient } from '@/utils/siweClient';
 import { ConnectKitProvider, getDefaultConfig } from 'connectkit';
 import type { AppProps } from 'next/app';
-import { WagmiConfig, createConfig } from 'wagmi';
+import { WagmiProvider, createConfig } from 'wagmi';
 
 const config = createConfig(
   getDefaultConfig({
-    alchemyId: process.env.NEXT_PUBLIC_ALCHEMY_KEY,
+    alchemyId: process.env.NEXT_PUBLIC_ALCHEMY_ID,
     walletConnectProjectId: process.env.NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID!,
     appName: 'My ConnectKit App',
   })
@@ -14,12 +14,12 @@ const config = createConfig(
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
-    <WagmiConfig config={config}>
+    <WagmiProvider config={config}>
       <siweClient.Provider>
         <ConnectKitProvider>
           <Component {...pageProps} />
         </ConnectKitProvider>
       </siweClient.Provider>
-    </WagmiConfig>
+    </WagmiProvider>
   );
 }