+++ examples/testbench/src/components/Web3Provider.tsx
@@ -1,0 +1,69 @@
+import { createElement, createContext, useContext, useState } from 'react';
+import { TestBenchProvider } from '../TestbenchProvider';
+
+import { getDefaultConfig, wallets } from 'connectkit';
+
+import { WagmiProvider, createConfig } from 'wagmi';
+import { defineChain, type Chain, http } from 'viem';
+
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+
+const avalanche: Chain = defineChain({
+  id: 43_114,
+  name: 'Avalanche',
+  nativeCurrency: {
+    decimals: 18,
+    name: 'Avalanche',
+    symbol: 'AVAX',
+  },
+  rpcUrls: {
+    default: { http: ['https://api.avax.network/ext/bc/C/rpc'] },
+  },
+  blockExplorers: {
+    default: { name: 'SnowTrace', url: 'https://snowtrace.io' },
+    snowtrace: { name: 'SnowTrace', url: 'https://snowtrace.io' },
+  },
+  testnet: false,
+});
+
+const ckConfig = getDefaultConfig({
+  chains: [avalanche],
+  transports: {
+    [avalanche.id]: http(avalanche.rpcUrls.default.http[0]),
+  },
+  appName: 'ConnectKit testbench',
+  appIcon: '/app.png',
+  walletConnectProjectId: process.env.NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID!,
+});
+const customConfig = {
+  ...ckConfig,
+  connectors: [wallets['rainbow'], ...(ckConfig.connectors ?? [])],
+};
+const config = createConfig(ckConfig);
+const queryClient = new QueryClient();
+
+type ContextValue = {};
+
+const Context = createContext<ContextValue | null>(null);
+
+export const Web3Provider = ({ children }: { children: React.ReactNode }) => {
+  return createElement(
+    Context.Provider,
+    { value: {} },
+    <WagmiProvider config={config}>
+      <QueryClientProvider client={queryClient}>
+        <TestBenchProvider
+        //customTheme={{ '--ck-font-family': 'monospace' }}
+        >
+          {children}
+        </TestBenchProvider>
+      </QueryClientProvider>
+    </WagmiProvider>
+  );
+};
+
+export const useWeb3 = () => {
+  const context = useContext(Context);
+  if (!context) throw Error('useWeb3 must be inside a Web3Provider.');
+  return context;
+};