+++ packages/connectkit/src/components/ConnectKit.tsx
@@ -19,7 +19,6 @@ import defaultTheme from '../styles/defaultTheme';
 import ConnectKitModal from '../components/ConnectModal';
 import { ThemeProvider } from 'styled-components';
 import { useThemeFont } from '../hooks/useGoogleFont';
-import { useNetwork } from 'wagmi';
 import { SIWEContext } from './../siwe';
 import { useChains } from '../hooks/useChains';
 import {
@@ -28,6 +27,8 @@ import {
 } from '../hooks/useConnectCallback';
 import { isFamily } from '../utils/wallets';
 import { useConnector } from '../hooks/useConnectors';
+import { WagmiContext, useAccount } from 'wagmi';
+import { Web3ContextProvider } from './contexts/web3';
 
 export const routes = {
   ONBOARDING: 'onboarding',
@@ -43,7 +44,6 @@ export const routes = {
 
 type Connector = {
   id: string;
-  name?: string;
 };
 type Error = string | React.ReactNode | null;
 
@@ -107,7 +107,7 @@ type ConnectKitProviderProps = {
   debugMode?: boolean;
 } & useConnectCallbackProps;
 
-export const ConnectKitProvider: React.FC<ConnectKitProviderProps> = ({
+export const ConnectKitProvider = ({
   children,
   theme = 'auto',
   mode = 'auto',
@@ -116,7 +116,12 @@ export const ConnectKitProvider: React.FC<ConnectKitProviderProps> = ({
   onConnect,
   onDisconnect,
   debugMode = false,
-}) => {
+}: ConnectKitProviderProps) => {
+  // ConnectKitProvider must be within a WagmiProvider
+  if (!React.useContext(WagmiContext)) {
+    throw Error('ConnectKitProvider must be within a WagmiProvider');
+  }
+
   // Only allow for mounting ConnectKitProvider once, so we avoid weird global
   // state collisions.
   if (React.useContext(Context)) {
@@ -131,6 +136,7 @@ export const ConnectKitProvider: React.FC<ConnectKitProviderProps> = ({
   });
 
   const chains = useChains();
+
   const injectedConnector = useConnector('injected');
 
   // Default config options
@@ -180,7 +186,6 @@ export const ConnectKitProvider: React.FC<ConnectKitProviderProps> = ({
   const [open, setOpen] = useState<boolean>(false);
   const [connector, setConnector] = useState<ContextValue['connector']>({
     id: '',
-    name: undefined,
   });
   const [route, setRoute] = useState<string>(routes.CONNECTORS);
   const [errorMessage, setErrorMessage] = useState<Error>('');
@@ -196,9 +201,12 @@ export const ConnectKitProvider: React.FC<ConnectKitProviderProps> = ({
   useEffect(() => setErrorMessage(null), [route, open]);
 
   // Check if chain is supported, elsewise redirect to switches page
-  const { chain } = useNetwork();
+  const { chain } = useAccount();
   useEffect(() => {
-    if (opts.enforceSupportedChains && chain?.unsupported) {
+    if (
+      opts.enforceSupportedChains &&
+      Boolean(chain && !chains.some((x) => x.id !== chain?.id))
+    ) {
       setOpen(true);
       setRoute(routes.SWITCHNETWORKS);
     }
@@ -250,15 +258,17 @@ export const ConnectKitProvider: React.FC<ConnectKitProviderProps> = ({
     Context.Provider,
     { value },
     <>
-      <ThemeProvider theme={defaultTheme}>
-        {children}
-        <ConnectKitModal
-          lang={ckLang}
-          theme={ckTheme}
-          mode={mode}
-          customTheme={ckCustomTheme}
-        />
-      </ThemeProvider>
+      <Web3ContextProvider>
+        <ThemeProvider theme={defaultTheme}>
+          {children}
+          <ConnectKitModal
+            lang={ckLang}
+            theme={ckTheme}
+            mode={mode}
+            customTheme={ckCustomTheme}
+          />
+        </ThemeProvider>
+      </Web3ContextProvider>
     </>
   );
 };