+++ packages/connectkit/src/components/ConnectModal/demo.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useRef, useState } from 'react';
-import { useAccount, useNetwork } from 'wagmi';
+import { useAccount } from 'wagmi';
 import { routes, useContext } from '../ConnectKit';
 import { CustomTheme, Languages, Theme, Mode } from '../../types';
 import Modal from '../Common/Modal';
@@ -20,6 +20,7 @@ import { ConnectKitThemeProvider } from '../ConnectKitThemeProvider/ConnectKitTh
 
 import styled from './../../styles/styled';
 import { keyframes } from 'styled-components';
+import { useChains } from '../../hooks/useChains';
 
 const dist = 8;
 const shake = keyframes`
@@ -97,12 +98,13 @@ const ConnectModal: React.FC<{
   onClose,
 }) => {
   const context = useContext();
-  const { isConnected } = useAccount();
-  const { chain } = useNetwork();
+  const { isConnected, chain } = useAccount();
+  const chains = useChains();
 
   //if chain is unsupported we enforce a "switch chain" prompt
   const closeable = !(
-    context.options?.enforceSupportedChains && chain?.unsupported
+    context.options?.enforceSupportedChains &&
+    Boolean(chain && !chains.some((x) => x.id !== chain?.id))
   );
 
   const showBackButton =