+++ packages/connectkit/src/components/Pages/SwitchNetworks/index.tsx
@@ -6,7 +6,7 @@ import {
   ModalBody,
 } from '../../Common/Modal/styles';
 import ChainSelectList from '../../Common/ChainSelectList';
-import { useAccount, useConnect, useDisconnect, useNetwork } from 'wagmi';
+import { useAccount, useConnect, useDisconnect } from 'wagmi';
 
 import useLocales from '../../../hooks/useLocales';
 
@@ -14,12 +14,13 @@ import Button from '../../Common/Button';
 import { DisconnectIcon } from '../../../assets/icons';
 import { isSafeConnector } from '../../../utils';
 import { OrDivider } from '../../Common/Modal';
+import { useChains } from '../../../hooks/useChains';
 
 const SwitchNetworks: React.FC = () => {
   const { reset } = useConnect();
   const { disconnect } = useDisconnect();
-  const { chain } = useNetwork();
-  const { connector } = useAccount();
+  const { connector, chain } = useAccount();
+  const chains = useChains();
 
   const locales = useLocales({});
 
@@ -31,7 +32,7 @@ const SwitchNetworks: React.FC = () => {
   return (
     <PageContent style={{ width: 278 }}>
       <ModalContent style={{ padding: 0, marginTop: -10 }}>
-        {chain?.unsupported && (
+        {Boolean(chain && !chains.some((x) => x.id !== chain?.id)) && (
           <ModalBody>
             {locales.warnings_chainUnsupported}{' '}
             {locales.warnings_chainUnsupportedResolve}
@@ -42,18 +43,19 @@ const SwitchNetworks: React.FC = () => {
           <ChainSelectList variant="secondary" />
         </div>
 
-        {chain?.unsupported && !isSafeConnector(connector?.id) && (
-          <div style={{ paddingTop: 12 }}>
-            <OrDivider />
-            <Button
-              icon={<DisconnectIcon />}
-              variant="secondary"
-              onClick={onDisconnect}
-            >
-              {locales.disconnect}
-            </Button>
-          </div>
-        )}
+        {Boolean(chain && !chains.some((x) => x.id !== chain?.id)) &&
+          !isSafeConnector(connector?.id) && (
+            <div style={{ paddingTop: 12 }}>
+              <OrDivider />
+              <Button
+                icon={<DisconnectIcon />}
+                variant="secondary"
+                onClick={onDisconnect}
+              >
+                {locales.disconnect}
+              </Button>
+            </div>
+          )}
       </ModalContent>
     </PageContent>
   );